+++ src/App.js
@@ -7,29 +7,31 @@ import {
 } from "react-router-dom"
 import './App.css';
 import stockData from "../data/stock-data"
+import Dashboard from "./dashboard"
 
 
 
 class App extends Component {
   render() {
-    let stocks = stockData.map((stock, index) => {
-      return (
-        <div key={index}>
-          <p>{stock.name}</p>
-          <p>{stock.symbol}</p>
-          <p>{stock.lastPrice}</p>
-        </div>
-      )
-    })
+
     return (
       <Router>
         <div>
         <p>Navbar</p>
           <nav>
-              <div><Link to="/home">Home</Link></div>
+              <div><Link to="/dashboard">Home</Link></div>
               <div><Link to="/search">Search</Link></div>
            </nav>
-           {stocks}
+           <main>
+            <Route
+              path="/dashboard"
+              render={() => {
+                return(
+                  <Dashboard />
+                )
+              }}
+            />
+           </main>
         </div>
       </Router>
     );
