--- src/home.js
@@ -1,14 +1,0 @@
-import React, {Component} from "react";
-
-
-class Home extends Component{
-	render() {
-
-		return (
-			<div>
-				<h1>All Stocks</h1>
-
-			</div>
-			)
-	}
-}
